+++ TREnvironmentEditor/Model/Types/Entities/EMMovePickupFunction.cs
@@ -121,7 +121,7 @@ public class EMMovePickupFunction : BaseEMFunction
         else
         {
             // Only look for the types we are interested in.
-            matchingEntities = entities.FindAll(e => Types.Contains(e.TypeID));
+            matchingEntities = entities.FindAll(e => Types.Contains((short)e.TypeID));
         }
 
         foreach (TR2Entity match in matchingEntities)
