+++ TRLevelControl/IO/TRLevelWriter.cs
@@ -138,7 +138,7 @@ public class TRLevelWriter : BinaryWriter
 
     public void Write(TR2Entity entity)
     {
-        Write(entity.TypeID);
+        Write((short)entity.TypeID);
         Write(entity.Room);
         Write(entity.X);
         Write(entity.Y);
